Add free standing functions to use TPM2 Crypto Backend with an ESYS_CONTEXT*
